# api/openapi.yml
openapi: 3.1.0
info:
  title: VoteChain Governance API
  version: 1.0.0
  description: REST API for VoteChain governance with stable contracts.
  contact:
    name: VoteChain Team
servers:
  - url: /api/v1
    description: API v1 (current stable)
paths:
  /proposals:
    get:
      summary: List proposals
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 50
        - name: state
          in: query
          schema:
            type: string
            enum: [active, passed, rejected, executed, cancelled]
      responses:
        '200':
          description: A page of proposals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalSummaryListResponse'
  /proposals/{id}:
    get:
      summary: Get full proposal details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Proposal detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalDetailResponse'
        '404':
          description: Proposal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /proposals/{id}/votes:
    get:
      summary: List votes for a proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Vote records for a proposal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteRecordListResponse'
        '404':
          description: Proposal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /voters/{address}/votes:
    get:
      summary: Get vote history for a voter
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Voter vote history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteRecordListResponse'
  /governance/stats:
    get:
      summary: Get governance health metrics
      responses:
        '200':
          description: Governance statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GovernanceStatsResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /metrics/cache:
    get:
      summary: Cache hit/miss counters
      responses:
        '200':
          description: Cache metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheMetricsResponse'
components:
  schemas:
    # ── Envelope ──────────────────────────────────────────────────────────────
    ApiError:
      type: object
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
          example: Proposal not found
      required: [code, message]
    ApiResponse:
      type: object
      description: Standard response envelope used by all endpoints.
      properties:
        data:
          description: The response payload, or null on error.
          nullable: true
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ApiError'
          nullable: true
          description: Non-null when the request failed.
        meta:
          type: object
          nullable: true
          additionalProperties: true
          description: Optional metadata (pagination, cache info, etc.).
      required: [data, errors, meta]
    # ── Domain schemas ─────────────────────────────────────────────────────────
    ProposalState:
      type: string
      enum: [active, passed, rejected, executed, cancelled]
    ProposalSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        state:
          $ref: '#/components/schemas/ProposalState'
        quorum:
          type: integer
        start_time:
          type: integer
        end_time:
          type: integer
      required: [id, title, state, quorum, start_time, end_time]
    ProposalDetail:
      type: object
      properties:
        id:
          type: integer
        proposer:
          type: string
        title:
          type: string
        description:
          type: string
        quorum:
          type: integer
        votes_yes:
          type: integer
        votes_no:
          type: integer
        votes_abstain:
          type: integer
        start_time:
          type: integer
        end_time:
          type: integer
        state:
          $ref: '#/components/schemas/ProposalState'
        execute_after:
          type: integer
      required: [id, proposer, title, description, quorum, votes_yes, votes_no, votes_abstain, start_time, end_time, state, execute_after]
    VoteChoice:
      type: string
      enum: [yes, no, abstain]
    VoteRecord:
      type: object
      properties:
        proposal_id:
          type: integer
        voter:
          type: string
        vote:
          $ref: '#/components/schemas/VoteChoice'
        weight:
          type: integer
      required: [proposal_id, voter, vote, weight]
    GovernanceStats:
      type: object
      properties:
        byState:
          type: object
          additionalProperties:
            type: integer
        participationOverTime:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              rate:
                type: number
            required: [date, rate]
        topVoters:
          type: array
          items:
            type: object
            properties:
              address:
                type: string
              total_weight:
                type: integer
            required: [address, total_weight]
        avgQuorumAchievement:
          type: number
      required: [byState, participationOverTime, topVoters, avgQuorumAchievement]
    CacheMetrics:
      type: object
      properties:
        hits:
          type: integer
        misses:
          type: integer
        hitRate:
          type: number
      required: [hits, misses, hitRate]
    # ── Typed envelope aliases ─────────────────────────────────────────────────
    ProposalSummaryListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ProposalSummary'
              nullable: true
    ProposalDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/ProposalDetail'
              nullable: true
    VoteRecordListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/VoteRecord'
              nullable: true
    GovernanceStatsResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/GovernanceStats'
              nullable: true
    CacheMetricsResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/CacheMetrics'
              nullable: true
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              nullable: true
              default: null
            errors:
              type: array
              items:
                $ref: '#/components/schemas/ApiError'
              minItems: 1