# backend/src/routes/proposals.ts
/**
 * Proposal routes with Redis caching applied.
 * On-chain data is fetched via the shared Stellar RPC service.
 */

import { Router, Request, Response } from "express";
import {
  cacheProposalList,
  cacheProposalItem,
  getCacheMetrics,
  invalidateProposalCache,
} from "../middleware/redisCache";
import { validate } from "../middleware/requestValidator";
import { sendSuccess } from "../middleware/response";

const router = Router();

// GET /proposals — cached 30 s
router.get(
  "/proposals",
  validate({
    query: {
      offset: { type: "integer", required: false, min: 0 },
      limit: { type: "integer", required: false, min: 1, max: 50 },
      state: { type: "string", required: false, enum: ["active", "passed", "rejected", "executed", "cancelled"] },
    },
  }),
  cacheProposalList,
  async (_req: Request, res: Response) => {
    // TODO: fetch from Stellar RPC / indexer
    const proposals: unknown[] = [];
    sendSuccess(res, proposals);
  }
);

// GET /proposals/:id — cached 10 s
router.get(
  "/proposals/:id",
  validate({
    params: {
      id: { type: "integer", required: true },
    },
  }),
  cacheProposalItem,
  async (req: Request, res: Response) => {
    const { id } = req.params;
    // TODO: fetch single proposal from Stellar RPC / indexer
    sendSuccess(res, { id });
  }
);

// GET /proposals/:id/votes — get votes for a proposal
router.get(
  "/proposals/:id/votes",
  validate({
    params: {
      id: { type: "integer", required: true },
    },
  }),
  async (req: Request, res: Response) => {
    const { id } = req.params;
    // TODO: fetch proposal votes from Stellar RPC / indexer
    sendSuccess(res, []);
  }
);

// POST /proposals/invalidate — called by the event indexer on new on-chain events
router.post(
  "/proposals/invalidate",
  validate({
    body: {
      id: { type: "string", required: false, min: 1, max: 64, pattern: /^[a-zA-Z0-9_-]+$/ },
    },
  }),
  async (req: Request, res: Response) => {
    const { id } = req.body as { id?: string };
    await invalidateProposalCache(id);
    res.json({ ok: true, invalidated: id ?? "list" });
  }
);

// GET /metrics/cache — exposes hit/miss counters
router.get("/metrics/cache", (_req: Request, res: Response) => {
  sendSuccess(res, getCacheMetrics());
});

export default router;